requirements.txt
# ── Google Cloud Vertex AI + Storage ─────────────────────────────────────────
google-cloud-aiplatform>=1.45.0
google-cloud-storage>=2.14.0

# ── Core data science ─────────────────────────────────────────────────────────
pandas>=2.1.0
numpy>=1.26.0
scikit-learn>=1.4.0
openpyxl>=3.1.0         # Excel read/write

# ── Visualization ─────────────────────────────────────────────────────────────
matplotlib>=3.8.0
seaborn>=0.13.0

# ── Config & env ─────────────────────────────────────────────────────────────
python-dotenv>=1.0.0
pyyaml>=6.0

# ── Video processing ──────────────────────────────────────────────────────────
moviepy>=1.0.3           # Video trimming
yt-dlp[default]>=2024.1  # YouTube download
# ffmpeg: system package — install via: sudo apt-get install -y ffmpeg

# ── Jupyter / notebook ────────────────────────────────────────────────────────
jupyter>=1.0.0
ipykernel>=6.29.0
